Create Thunder base schema and move common types into it
